+++ src/connect/component.test.js
@@ -16,7 +16,16 @@ vi.mock("@paypal/sdk-client/src", () => {
     getClientID: vi.fn(() => "mock-client-id"),
     getClientMetadataID: vi.fn(() => "mock-cmid"),
     getUserIDToken: vi.fn(() => "mock-uid"),
-    getLogger: vi.fn(() => ({ metric: vi.fn(), error: vi.fn() })),
+    getDebug: vi.fn(() => false),
+    getLogger: vi.fn(() => ({
+      metric: vi.fn().mockReturnThis(),
+      error: vi.fn().mockReturnThis(),
+      track: vi.fn().mockReturnThis(),
+      flush: vi.fn().mockReturnThis(),
+    })),
+    getEnv: vi.fn(),
+    getCSPNonce: vi.fn(),
+    loadFraudnet: vi.fn(() => ({ collect: vi.fn() })),
   };
 });
 
@@ -60,8 +69,9 @@ describe("getConnectComponent: returns ConnectComponent", () => {
       platformOptions: {
         platform: "PPCP",
         clientID: "mock-client-id",
-        clientMetadataID: "mock-cmid",
+        clientMetadataId: "mock-cmid",
         userIdToken: "mock-uid",
+        fraudnet: expect.any(Function),
       },
     });
     expect(sendCountMetric).toBeCalledTimes(2);
@@ -85,5 +95,15 @@ describe("getConnectComponent: returns ConnectComponent", () => {
       expectedError
     );
     expect(sendCountMetric).toBeCalledTimes(2);
+  });
+
+  test("minified is set according to debug value", async () => {
+    await getConnectComponent(mockProps);
+    expect(loadAxo).toHaveBeenCalledWith({
+      minified: true,
+      btSdkVersion: "3.97.3-connect-alpha.6.1",
+      metadata: undefined,
+      platform: "PPCP",
+    });
   });
 });