+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## <small>5.0.291 (2023-12-15)</small>
+
+* feat: provide fraudnet preload functionality and wire up connect with it (#2287) ([fac18df](https://github.com/paypal/paypal-checkout-components/commit/fac18df)), closes [#2287](https://github.com/paypal/paypal-checkout-components/issues/2287)
+
+
+
+## <small>5.0.290 (2023-12-15)</small>
+
+* add debug support for axo (#2292) ([d858f23](https://github.com/paypal/paypal-checkout-components/commit/d858f23)), closes [#2292](https://github.com/paypal/paypal-checkout-components/issues/2292)
+
+
+
 ## <small>5.0.289 (2023-11-30)</small>
 
 * fix: update to connect loader 1.1.1 (#2282) ([77b7dc0](https://github.com/paypal/paypal-checkout-components/commit/77b7dc0)), closes [#2282](https://github.com/paypal/paypal-checkout-components/issues/2282)